:truck: Moved hero to Modules folder
