+++ src/Common/Components/Modules/Hero.scss
